Moved the unittest files into the unittests module.
